pyproject.toml
# pyproject.toml
# =======================================

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bit_collision_free_MF"
dynamic = ["version"]                      
description = "Generate molecular fingerprints with guaranteed collision‑free bits."
readme = "README.md"
license = {file = "LICENSE"}              
requires-python = ">=3.9"

authors = [
  {name = "Shifa Zhong", email = "sfzhong@tongji.edu.cn"},
  {name = "Jibai Li",   email = "51263903065@stu.ecnu.edu.cn"},
]

dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "rdkit-pypi",
]

classifiers = [
  "License :: OSI Approved :: MIT License",
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Chemistry",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Shifa-Zhong/bit_collision_free_MF"
Documentation = "https://github.com/Shifa-Zhong/bit_collision_free_MF#readme"

# ---------------------------------------
# setuptools 额外配置
[tool.setuptools.dynamic]
version = {attr = "bit_collision_free_MF.__version__"}  

[tool.setuptools.packages.find]
where = ["."]
include = ["bit_collision_free_MF*"]
